feat(user): 添加用户注册功能并优化登录页面
- 新增用户注册页面和相关功能
-改进登录页面，增加注册链接
- 实现用户密码加密存储
- 调整全局异常处理，显示具体错误信息
- 优化全局变量注入，移除重复代码

+++ src/main/java/org/hao/compiler/entity/User.java
@@ -109,13 +109,13 @@ public class User extends Model<User> implements Serializable {
      * 账号状态：0禁用，1启用
      */
     @Schema(description = "账号状态：0禁用，1启用")
-    private Byte status;
+    private Integer status;
 
     /**
      * 是否删除：0未删，1已删
      */
     @Schema(description = "是否删除：0未删，1已删")
-    private Byte isDeleted;
+    private Integer isDeleted;
 
     /**
      * 创建人ID或用户名
